Now any command understand defaultMenuIcon which by default ask uses icon name from class side (defaultMenuIconName). They both can return nil which is default value.
Also additional method is added for the cases when real icon instance should be ensured: #secureMenuIcon. In case if missin icon it returnes error icon.

And CmdCommandActivator is now understands menuItemIcon.
All commands are adopted to this simplification.

And it is merged with Esteban spotter menu. But this part needs to be repackaged