Browser annotation contexts are refactored. There are new composable context and they are now named with prefix expected:
- ClyExpectedBrowserContext
Simply use class of browser in annotations

- ClyExpectedQueryContext
Simply use class of query to active commands and tools in context of view which executes given kind of query.

- ClyExpectedSelectionContext
It is renamed ClyBrowserItemContext. It represent selected items of concrete type

- ClyExpectedQueryItemContext
It allow activate browser command and tools inc context of views which query retrieves particular kind of items. It not depends on kind of query.
To use it send converting message asQueryItemTypeContext to the type class.

Context is only introduced and they are not used yet.