Better selectClass: when full hierarchy mode. Now it force hierarchy of requested class when It is not in current hierarchy. - switchToFullClassHierarchyOf: is added to full browser
#selectedTextInterval is added to ClyTextEditorContext.
And SycExtractMethod uses it instead of selected source node interval which is not valid in case of code extraction.

MethodCreationTool is fixed to be not selected after method creation which happens in some cases when this tool is in desired selected tabs.

context menu building is moved to base menu activation class. It is now clear that we can ask any kind of menu to build context menu morph. It is not depends on type of activation. It can be needed in some scenarios for any type of menu activation. IT is also rename to #buildContextMenuFor:inContext:.

Fix for ClassDefinitionEditorTool which does not force class selection when it was editing of current class. In hierarhcy mode this editing can remove class from class view list. So we should always ask browser to select edited class.